requirements.txt
crewai
crewai_tools
load_dotenv
fastapi
uvicorn
PyPDF2
pandas
python-docx
pdfplumber
python-pptx
SpeechRecognition
moviepy
openpyxl
pydub
beautifulsoup4
markdown
pytube
wave
psycopg2
Flask 
requests
chardet